fixed minor issues

+++ src/main/java/com/github/rshtishi/Exercise9.java
@@ -6,7 +6,7 @@ public class Exercise9 {
         private double x;
         private double y;
 
-        public Point2D(int x, int y) {
+        public Point2D(double x, double y) {
             this.x = x;
             this.y = y;
         }
